Allow selection of text in a descendant element
The target of a selection event can be designated by adding the
'select-text-target' directive to a descendant. The contents of
that element will be selected instead of the element that was
clicked on.

Removed dependency on jQuery.

Using a linked directive and a shared service instead of using
jQuery's find function to get the selection target element.